Move JobTestCase to test.python.ibmq
