+++ test/python/ibmq/test_ibmqjob_states.py
@@ -17,7 +17,7 @@ from qiskit.providers.jobstatus import JobStatus
 from qiskit.providers.ibmq.ibmqjob import IBMQJobPreQobj, IBMQJob, API_FINAL_STATES
 from qiskit.providers.ibmq.api import ApiError
 from qiskit.providers import JobError, JobTimeoutError
-from ..common import JobTestCase
+from .jobtestcase import JobTestCase
 from .._mockutils import new_fake_qobj, FakeBackend
